(PC-29248)[PRO] refactor: extract orejime logic into a hook
