+++ pro/src/app/App/App.tsx
@@ -1,5 +1,5 @@
 import { setUser as setSentryUser } from '@sentry/browser'
-import React, { useEffect, useState } from 'react'
+import React, { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { Navigate, Outlet, useLocation } from 'react-router-dom'
 import { SWRConfig } from 'swr'
@@ -14,8 +14,8 @@ import useIsNewInterfaceActive from 'hooks/useIsNewInterfaceActive'
 import useNotification from 'hooks/useNotification'
 import { updateUser } from 'store/user/reducer'
 import { selectCurrentUser } from 'store/user/selectors'
-import { Consents, initCookieConsent } from 'utils/cookieConsentModal'
 
+import { useOrejime } from './analytics/orejime'
 import useFocus from './hook/useFocus'
 import { useLoadFeatureFlags } from './hook/useLoadFeatureFlags'
 import useLogNavigation from './hook/useLogNavigation'
@@ -27,10 +27,9 @@ export const App = (): JSX.Element | null => {
   const isBeamerEnabled = useActiveFeature('ENABLE_BEAMER')
   const location = useLocation()
   const currentUser = useSelector(selectCurrentUser)
-  const [consentedToFirebase, setConsentedToFirebase] = useState(false)
-  const [consentedToBeamer, setConsentedToBeamer] = useState(false)
   const dispatch = useDispatch()
   const notify = useNotification()
+  const { consentedToBeamer, consentedToFirebase } = useOrejime()
 
   const isNewInterfaceActive = useIsNewInterfaceActive()
 
@@ -40,41 +39,6 @@ export const App = (): JSX.Element | null => {
       isNewInterfaceActive ? 'blue' : 'pink'
     )
   }, [isNewInterfaceActive])
-
-  useEffect(() => {
-    // Initialize cookie consent modal
-    if (location.pathname.indexOf('/adage-iframe') === -1) {
-      setTimeout(() => {
-        const orejime = initCookieConsent()
-        // Set the consent on consent change
-        orejime.internals.manager.watch({
-          update: ({
-            consents,
-          }: {
-            consents: { [key in Consents]: boolean }
-          }) => {
-            setConsentedToFirebase(consents[Consents.FIREBASE])
-            setConsentedToBeamer(consents[Consents.BEAMER])
-          },
-        })
-        setConsentedToFirebase(
-          orejime.internals.manager.consents[Consents.FIREBASE]
-        )
-        setConsentedToBeamer(
-          orejime.internals.manager.consents[Consents.BEAMER]
-        )
-
-        // We force the banner to be displayed again if the cookie was deleted somehow
-        if (!document.cookie.includes('orejime')) {
-          orejime.internals.manager.confirmed = false
-          initCookieConsent()
-        }
-      })
-    } else {
-      setConsentedToFirebase(false)
-      setConsentedToBeamer(false)
-    }
-  }, [location.pathname])
 
   useConfigureFirebase({
     currentUserId: currentUser?.id.toString(),
