+++ pro/src/pages/CookiesFooter/CookiesFooter.tsx
@@ -1,11 +1,11 @@
 import cn from 'classnames'
 import React from 'react'
 
+import { orejime } from 'app/App/analytics/orejime'
 import useIsNewInterfaceActive from 'hooks/useIsNewInterfaceActive'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonLink } from 'ui-kit/Button/ButtonLink'
 import { ButtonVariant } from 'ui-kit/Button/types'
-import { initCookieConsent } from 'utils/cookieConsentModal'
 
 import styles from './CookiesFooter.module.scss'
 
@@ -50,7 +50,7 @@ const CookiesFooter = ({ className }: { className?: string }) => {
             <Button
               variant={ButtonVariant.QUATERNARY}
               className={styles['cookies-footer-link']}
-              onClick={() => initCookieConsent().show()}
+              onClick={() => orejime.show()}
             >
               Gestion des cookies
             </Button>
